Update patch set 3
Patch Set 3:

> For "classic" dev mode, I think it would make sense for log
 > messages to automatically go the the dev mode console. However,
 > it's not quite so simple to set up.
 > 
 > In Super Dev Mode, GWT.log() goes to the JavaScript console, so if
 > the console logger calls GWT.log() and then logs to the console
 > itself, the message will be logged twice.
 > 
 > We would also need to make sure we don't log messages twice in
 > GWTTestCase. I think if you're using GWTTestCase using "classic"
 > dev mode, messages logged with GWT.log() already go to stdout or
 > stderr?
 > 

It goes to TreeLogger which I think only outputs to dev-mode window.
I disabled the dev-mode handler for GWTTestCase anyway so it shouldn't be a problem even it goes to stdout.

 > Long-term, I think we should keep GWT.log() since it has a
 > different purpose than java.util.Logging. The idea behind GWT.log()
 > is to have deterministic behavior so that it can be relied on for
 > debugging during development. In "classic" Dev Mode, it always goes
 > to the dev mode console. In Super Dev Mode, it always goes to the
 > JavaScript console. In production, it's always compiled out.
 > 

That sounds exactly like a specific configuration to your logging which is able to provide the same semantic (e.g. info level log goes to console only in dev mode and compiled out in production). We just need more meaningful defaults for each environment (ie. dev, testing, production).

The additional benefit is if you are hunting a bug for specific environment, you can change your configuration and get additional debug information. That's something java developers already accustomed to; They have both logger and system.out.println but they will not submit the code with println and it will not compile away.

 > Configurable logging also has its place, but we don't need two
 > configurable logging systems. I think java.util.Logging meets the
 > need for configurable logging well. Using GWT.log() is more like
 > debugging using print statements, where making it configurable
 > would make it less useful.

A more appropriate analogy to print statements is a Console abstraction that people could use is to print temporary debugging statements; such print statements do not compile away and you basically don't submit them if you don't want them for production.

Patch-set: 3