Fix $permProps to work with -XenableClosureCompiler
Closure was pruning the $permProps variable, resulting in a
ReferenceError on the last line of the module suffix.

Prevented Closure from removing it by switching to a global
assignment and adding an extern declaration.

Prevented the ReferenceError (even if $permProps is not set,
as happens with devmode.js) by declaring the variable in the
module prefix.

When not using CrossSiteIframeLinker, the module prefix won't be
there and $permProps won't be declared. But that's okay. It will
be a global on "window" like everything else.

Change-Id: I2200c01e79fb229664584f80931ae5452a8cbf84
Review-Link: https://gwt-review.googlesource.com/#/c/7860/