Update patch set 3
Patch Set 3: Code-Review+1

(1 comment)

Thanks for detailed description. It took me for a while to understand what is going on but now it makes sense. 

It sounds like this is basics of the problem should be easy to reproduce with test-case by setting and getting scrollTop/scrollLeft. Can you take a look at that?

Patch-set: 3
Label: Code-Review=+1