Update patch set 4
Patch Set 4:

One other alternative is to keep using Object.keys but get rid of the for loops and copy for constructing the iterator. In that case we wouldn't need to worry about the slowness because that is going to be definitely faster in all modern browsers.
However there is a catch: to implement the iterators without copy, we should start throwing ConcurrentModificationException which is going to very likely break some projects as we were not doing that before. I'll take look into that and see how breaking it is internally for Google.

Patch-set: 4
Label: Verified=0