Update patch set 1
Patch Set 1:

Ok, the comment was misleading.

The hash is already encoded. The hash 'characters' that shows up in the middle of the hash will be started to get encoded.

Also I think we don't need to refer '2.7' in the javadoc, just the description of the behavior is enough..

Patch-set: 1