Update patch set 1
Patch Set 1:

Goktug is right. I've been misunderstanding this change all along because I somehow didn't see the call to URL.encode() in the old version.

I don't think there was any workaround for properly encoding the hash character before this change, because if you tried to do it then the % would get double-encoded. The only workaround was to not use a hash character. So I guess there isn't anything to warn about. My apologies for making a bigger deal out of this than it actually is.

Patch-set: 1