Update patch set 2
Patch Set 2:

Maybe we could make GWT.jsni less "magic"?

E.g.

 class JSNI {
   static <T> T expr(String, Object... args)
   static int intExpr(String, Object... args)
   static bool boolExp(String, Object... args)
   …
   static void stmt(String, Object...)
 }

JSNI.expr() could then fail to compile if the value is not a JS expression (e.g. contains a ";")

And, maybe more importantly, the returned value is not magically "not boxed" depending on the context.

Patch-set: 2
Label: Verified=0