Update patch set 3
Patch Set 3:

I just pushed another patch that implements the unboxing magic for parameters, but does not magically unbox return types.

I instead added GWT.jsniInt/jsniNumber/jsniBool/etc.

The reason is, I looked at magically removing the unboxing of the return type that the compiler inserts but it would make the GwtAstBuilder code way way more complex and potentially brittle, so I'd rather not go there.

I also want to add one more bit of magic, namely this:

GWT.jsni("$0()", () -> alert('Hello'));

That is, passing in lambdas or single-method-interfaces instead passes in a JSNI reference to the function bound to the this pointer of the reference, that is


GWT.jsni("$0()", new Runnable() { void run() { alert("hello"); } })

really does this:

public static native void jsni$0(JavaScriptObject $0) /*-{
   $0();
}-*/;

And at the call site:

jsni$0(($tmp = new Runnable$1(), ($tmp.@Runnable$1::run()).bind($tmp))

That is, it takes a JSNI ref to the ::run method, and binds the function to the instance, so that it is invokable in JS.

Use case:

GWT.jsni("$doc.body.onLoad = $0", () -> alert('Hello'));

Patch-set: 3
Label: Verified=0