Remove use of prebuilt requestfactory-apt.jar
The current setup needed requestfactory-apt.jar in GWT_TOOLS to be kept
up-to-date after each change to RF annotation processing, which proved to
be too much burden.

This change instead explicitly calls the RF annotation processor when
compiling tests. As a side-effect, it means that in case we later need
another annotation processor, it'll have to be explicitly called too, as
discovery is no longer used. We expect to have changed the build system
before then, which would allow us to use prebuilt JARs and go back to rely
on discovery for annotation processing.

Change-Id: I36f3c2efabd9a24bc1ba18e6961f4dd4f12514da