gwt-dev: Don't package jetty in maven artifact
Jetty is packaged in gwt-dev.jar for convenience only, to make DevMode,
SuperDevMode and GWT-JUnit work.

Unfortunately it is totally outdated and can and would collide with more
recent Jetty versions that are used by projects that consume gwt-dev.jar
artifact. More then that, because GWT currently depends on HtmlUnit and
HtmlUnit depends on outdated Jetty WebSockets, and Jetty WebSockets was
substantialy redesigned and splitted into 1000 libraries in most recent
Jetty version, the probability that HtmlUnit would upgrade to recent
Jetty version is more or less zero.

That means that GWT is going to ship outdated version of Jetty forever.

So the projects that depend on GWT and on most recent Jetty version have
a pain of messing around with two incompatible versions of the Jetty
library in their classpath.

To reduce the pain stop to package Jetty in gwt-dev.jar maven artifact.
Strip Jetty during upload to Central and add the dependency to it in
pom.xml instead.  That is the same way it's done for org.json package
(for license reasons though).

[1] https://gerrit.googlesource.com/gerrit/+/master/lib/gwt/BUCK

Bag: issue 8747
Change-Id: I9a0d321ad2981f2ec48f70d24a030fe387e5522d