gwt-dev: Don't package jetty in maven artifact
Jetty is packaged in gwt-dev.jar for convinience only, to make DevMode
and SuperDevMode work.

Unfortuntaly it is totally outdated and can and would collide with more
recent Jetty versions that are used by project that consume gwt-dev.jar
artifact.

Instead iof delegating the job of stripping the outdated Jetty version
from gwt-dev.jar before using [1] and after downloading it, strip it
from the maven artifact and add the dependency to it in pom.xml. That
is the same way it's fone for org.json packaging.

[1] https://gerrit.googlesource.com/gerrit/+/master/lib/gwt/BUCK

Bug: issue 8747
Change-Id: I9a0d321ad2981f2ec48f70d24a030fe387e5522d