Bump Jetty version to 9.2
The Jetty 9.2 quick start mechanism provides almost an order of
magnitude improvement in start time [1].

Beside that projects that depend on both GWT and most recent Jetty
version have to mess around with stripping outdated Jetty version from
gwt-dev, download outdated Jetty again as aggregate version to satisfy
codeserver dependency and make SDM work. Not to mention the fun fixing
the collision in Eclipse's classpath between current and outdated Jetty
versions. See the whole story here [2].

Jetty websocket layer was substantially redesigned and split into
multiple different JARs. Because GWT is not using websocket it doesn't
make any sense to package it in gwt-dev JAR: it is discontinued. If it's
needed, then it should be consumed directly and not through gwt-dev.jar.

[1] https://webtide.com/jetty-9-quick-start
[2] https://groups.google.com/d/topic/repo-discuss/YrFmmf3Y2Ro

Bug: issue 8747
Change-Id: Id1d338a01f4fc72b2adf2f43b57aeddfb06cfec1