Update patch set 5
Patch Set 5:

I didn't update the original doc drafting the GWT.jsni (actually proposed to rename it Js.call instead) but I have diverged from my initial thoughts.

What I really want is easily testable and potentially JRE emulatable JSNI. For that we shouldn't rely on the knowledge of the call-site. That helps testability and also simplifies the semantics.

So to avoid relying on call-site knowledge:

1. We should not add magic "this" context to GWT.jsni. So GWT.jsni should always translate into static methods. 'this' still could be passed as parameter if needed.

2. We should ditch relying on type inference for return types.

So my new proposal is more like following:

  // No returns
  Js.call("$wnd.alert('Hello')")

  // Returns an object
  Document doc = Js.call("$doc).as(Document.class);

  // Returns primitive
  int left = Js.call("$doc.scrollLeft").asInt();


So I know this is little bit more verbose and might cause more work on the compiler as there is an intermediate type now but I think this could be helpful in the long run.

The other proposal for indexed types;
I actually responded to that earlier in the doc; you might have missed that. I think we already have a good story for that with IsArrayLike[1]. Also developers can choose to use 'default' methods to add the same functionality but I would still prefer them to use IsArrayLike instead for keeping API compatibility.

[1] https://docs.google.com/a/google.com/document/d/1tir74SB-ZWrs-gQ8w-lOEV3oMY6u6lF2MmNivDEihZ4/view#bookmark=id.hwiq26rs2fja

Patch-set: 5