Update patch set 5
Patch Set 5:

>What I really want is easily testable and potentially JRE >emulatable JSNI. For that we shouldn't rely on the >knowledge of the call-site. That helps testability and also >simplifies the semantics.

How do you plan to make var-args work? I really don't think we should sacrifice our JS semantics to work in the JRE again. I don't think Nashorn really changes the equation anymore than HtmlUnit+Rhino did. We went down that rabbit hole once already with DevMode and bytecode rewriting everywhere. One of the benefits of the new JsInterop model is the freedom to assume the code won't work in DevMode/JRE -- that we are running pure JS.  If we can make it work fine, but I don't constrain interop by forcing it to be defined in a way that it has to work in JRE mode. Both DevMode and Dart have trouble by making assumptions of code running in two heaps with different semantics. 

>So to avoid relying on call-site knowledge:
>1. We should not add magic "this" context to GWT.jsni. So GWT.jsni should always translate into static >methods. 'this' still could be passed as parameter if needed.

It's not really magic, the compiler does nothing special to make this work. "It just works".  GWT.jsni puts a method on the same class it occurs in. I'm not particularly religious about this one, but it seems people can get confused both ways.

>2. We should ditch relying on type inference for return types.
I already ditched it, in favor of jsniInt, jsniBool, etc.

>  Document doc = Js.call("$doc).as(Document.class);

This is unneeded as Java generics already give you this for free.

>  // Returns primitive
>  int left = Js.call("$doc.scrollLeft").asInt();
>So I know this is little bit more verbose and might cause more work on the compiler as there is an >intermediate type now but I think this could be helpful in the long run.

I don't see how this is a win over Js.callInt, Js.callBool. The compiler would have to do even more magic, because it would have to eliminate a non-existent intermediate object that serves only as a cast, it's also more verbose. Essentially, you will have to pretend that the compiler returns some interface with the 'as' method on it, and then intercept and unwrap the JMethodCall and rewrite the expression. This doesn't seem to be a huge win over just having a few differently named methods.

>The other proposal for indexed types; I actually responded to that earlier in the doc; you might have >missed that. I think we already have a good story for that with IsArrayLike[1]. Also developers can

I don't think we can assume Java 8 only. Also, Js.set() would also depend on the compiler  to make efficient, and it requires developers to type more code. Declaring a property indexed very closely follows the WebIDL spec that Javascript already implements. If you don't want indexed properties, than why even want @JsProperty at all? You can also implement that with imperative calls. There are other reasons you want this declarative, and that's because we will need to add other attributes to @JsProperty later anyway, like enumerable, configurable, and writable. JS properties have a notion of whether or not they are read only, something you can't express via imperative setters alone. The full implementation of @JsProperty will never work correctly with Object.defineProperty() semantics for concrete subclasses.

Patch-set: 5