Make polymorphic dispatch work in v8.
Classes that override methods from parent classes did
not get optimized in v8, due to a problem with the
way hidden classes are built up in v8.

V8 requires prototypes of different classes to differ
in the set of properties in order to perform some
optimizations. This patch makes every prototype have a
unique field for that purpose.

Change-Id: I8ef1c117974e4bce9281194779a832505d9799b0