Update patch set 2
Patch Set 2:

Do we really need 2 such configurations?

If -strict.xml is used for presubmit, it'll mark the changes with Verify:-1 if you fail to comply to the stricter checks; so why not have them as errors too in the Ant build and in Eclipse?

OK for adding line length (I think we once bumped up the limit to 110c, which would explain the few offending files I guess) and trailing spaces checks though.

Patch-set: 2