Update patch set 1
Patch Set 1:

While I'm all for removing this run style and its dependency on RMIC, I think we should first deprecate it, ship it as deprecated in the upcoming 2.7, and only then remove it completely.

The reason is that it's documented at http://www.gwtproject.org/doc/latest/DevGuideTestingRemoteTesting.html, with a warning for the selenium run style about Internet Explorer suggesting to use the remoteweb run style.

We're lucky enough that 2.7 is only a few weeks away so we'll be able to apply this change real soon.

In the mean time, we should IMO update RunStyleRemoteWeb to print a deprecation warning telling users to migrate to the selenium run style.

Patch-set: 1