Experimental patch testing unboxed primitives
Not for submit or review. I'm using this to experiment with approaches to unboxing,
and testing internally on apps. This is not neccessarily how it will finally be done.

1. After all optimizations have finished, replace all calls to BoxType.valueOf(x) with x,
   and replace all calls to boxTypeVar.<primitiveType>value() with boxTypeRef.

2. Disable same-parameter value optimization since it will end up altering stuff like
"new Double(12)" into "new Double" with 12-inlined into the ctor, which makes the
unboxing pass break.

3. Modify Devirtualizer to handle devirtualized calls to Object, Comparable,
and java.lang.Number methods (all mapped to Double versions)

4. All field references to BoxType.value where BoxType is a Number subtype are
replaced with the qualifier. This handles calls to static methods and inlined
expressions.

5. Implement Cast Checks for number subtypes (instanceof Double => typeof(x) === 'number')

Not handled yet:
Does deal with Boolean, Long, other subtypes of Number, like Short/Integer/Float/Short/etc
correctly, but those are easy to fix.

All hidden behind -XdisableAutoboxing, should be used with -XdisableCastChecking.

Change-Id: I3518fcb1fe97254c8d6022946be92bd6a5954d5e