Update patch set 1
Patch Set 1: Code-Review+2

This will compile but will fail on the runtime. Anyway, yesterday Daniel told me that we are doing the release in jdk6 but not running the compiler tests so I think this will solve the problem for the purposes of the release.

Patch-set: 1
Reviewer: Gerrit User 5110 <5110@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+2