Update patch set 1
Patch Set 1:

I don't think it's needed, but it's also harmless (and makes it clear, along with the comments) that the file is not (meant to be) processed.

I have an incentive of not touching the file, as the change is not needed (and to avoid any issue when people change the build scripts), but I nevertheless won't -1 the change.

Patch-set: 1
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0