Update patch set 1
Patch Set 1:

> * possibly exclude "war/" instead of listing all the "war/"

Oh, right, there are two "war/" directories checked in under eclipse/, which is why it was user/war.

One is for LogExample, which looks like a mistake, because the files (LogExample.css and LogExample.html) are in the non-eclipse location as well.

The other is for code-museum. Which, when I open it in Eclipse, has several compile errors (TreeItem cstr missing) and hasn't been touched, except for once by Daniel doing some housekeeping in 4 years.

Maybe we could just remove the code-museum all together?

Patch-set: 1