Update patch set 4
Patch Set 4:

> > Have you tried https://github.com/tbroyer/gwt-sandbox/tree/gradle-build
 > > ?
 > >
 > > I don't use "gradle idea" as IDEA provides a better integration
 > > (IMO) by just "importing a gradle project".
 > >
 > > Also, my branch actually builds GWT ;-)
 > 
 > Why Gradle and not Buck?

This is a good question ;-)

I, for one, wouldn't be against Buck; and I'm sure Google would like it too.

One reason we want to switch to another build tool is IDE integration: make it easy for anyone to jump in and contribute. Ideally, it shouldn't be harder than "git clone" and import into your IDE.

We first looked at Maven, but I gave up: Maven is too directive. It's not "convention over configuration", it's "conventions or you're doomed". Not to mention that all dependencies must FIRST be deployed to a Maven repo. That was just too much work and too much hassle to be worth it (and believe me, I tried hard, for almost a year!).

Gradle is gaining momentum so it's likely that some contributors already use it; IDE integration is also improving every day (particularly in IDEA thanks to Google with Android Studio), and Gradle has good support for Eclipse and IDEA out of the box (without using the IDE's own support for Gradle; similar to "mvn eclipse:eclipse" vs. M2Eclipse). Last, but not least, Gradle makes it easy to consume dependencies coming from JCenter or Central (or any similar repo); this makes it easier to upgrade dependencies and test new things.

Buck, on the other hand, requires cloning and building it (with Ant!) before you can build GWT. IDE integration is rather "poor" (you hand-wrote an Eclipse one for Gerrit; and I'm still not sure how IDEA integration actually works). And consuming external dependencies requires extra work too (even with bucklets, you still have to manually re-build the transitive dependencies in Buck; there's absolutely no tooling).

If it were just me, I'd probably go with Buck+bucklets and invest some time tackling the external deps issue by adding some tooling (point a tool to a pom.xml or ivy.xml and let it generate the BUCK file for you; such a tool would tell you when several deps transitively depend on the same third dep with different versions, etc.) and help enhance the IDE integration; but when I blogged about Buck last year (just because I think it's a good tool), people commented that they hoped we wouldn't use it for GWT, and we should just use Gradle.

Patch-set: 4