Update patch set 2
Patch Set 2:

> Hi Brian/Goktug,
 > 
 > So, I added an "rpc.serializeFinalFields=audit" that will print an
 > ERROR for every final field found.
 > 
 > However, two things:
 > 
 > 1) The ERROR doesn't fail the build, which I could fix, but
 > 2) The ERROR output of "hey, here's a final field you should look
 > at" is basically 100% identical to the WARN output of "hey, here's
 > a final field I'm skipping".
 > 
 > To me, this means we can drop the "audit" flag, and just say:
 > 
 > Existing users will get serializeFinalFields=false (or false_nowarn
 > if they're using the old suppress flag), so no change in behavior.
 > 
 > If a user wants to try serializeFinalFields, they can change the
 > value from false_nowarn (unset the legacy setting if they want) to
 > just false, and look at the warnings.
 > 
 > If they fix (add transient) or okay each field that had a warning,
 > then they're ready to flip the flag over to serializeFinalFields=true.

IMHO if error logs appear the build should also fail otherwise its not an error. 

So either keep the "audit" flag and let the build fail or remove it and stick to warnings. The "audit" flag might be useful for CI builds and in that case it might be a good idea to make it possible to override the property defined in GWT module using a system property (if that's already possible then fine, haven't looked through all the code yet).

Patch-set: 2
Label: Verified=0