Makes sure IE11 is never matched to older IEs
We have seen instances of IE11 hitting our servers with MSIE in UA
strings and have documentMode > 10. We are not 100% sure but this is
more likely an alteration of UA string with a plugin or user itself.

As we know that regardless of the UA string if documentMode > 10,
browser will not work on an older IE permuations due to missing
createStyleSheet method, so it is simple to make sure that never
happens.

Change-Id: I5d43630943c7c6084274bb5853884eccd0894901