Makes sure IE11 is not matched for older permutations
We have seen instances of IE11 hitting our servers with MSIE string
in UA and have documentMode > 10. We are not %100 sure but this is
more likely and alteration of UA with a plugin or the user customization.

As we know that regardless of the UA string if documentMode > 10,
browser will not work on older IE permuations due to missing
createStyleSheet method on document, so it is simple to make sure
that never happens.

Change-Id: I5d43630943c7c6084274bb5853884eccd0894901