Update patch set 3
Patch Set 3:

Okay, patch set implements the "rpc.serializeFinalFields=audit" functionality.

Because the "check the flag" logic was in a static helper method that was itself invoked by ~2 levels of other static helper methods, I added a semi-hacky "out" parameter, a boolean[], that allows callers to check whether the build should be failed.

I think this is cleaner than adding a special GeneratorContext wrapper just for GWT-RPC, and smaller than adding support for this in GeneratorContext proper.

It was also a fairly small change, as only the 1 caller who cares about checking the "should I fail the build now?" flag (the call site directly in the generator main path, and not the various other call sites), has to pass/care about the flag. The others can just pass null.

Since the "rpc.serializeFinalFields=audit" now really does fail the build, I could not add an integration-level test (at least based on copy/pasting the other FinalFieldXxx tests), because the test's GWT compile would fail.

I'm sure there are ways of testing a "GWT compile that should fail"...I would technically look in to that if given a pointer, but I'm also pretty tempted to just leave that flag behavior for manual verification at this point. Given it's a temporary/nice-to-have feature, and not a core part of the patch.

Patch-set: 3
Label: Verified=0