Update patch set 4
Patch Set 4:

> That makes flipping the flag harder since it's not just a compiler flag flip, it's also a server-side flag flip.

True. The current patch handles that, for StandardSerializationPolicy, by writing the compile-flag into the server-config, which I believe is fine because that's the point of the policy file anyway (recording what the compiler knows for the server to use later).

For other impls of SerializationPolicy, I'm fine saying that rpc.serializeFinalFields=true is unsupported until you update your custom SerializationPolicy impl to somehow know about the config setting (or even hardcode it to true if the app will always use rpc.serializeFinalFields=true going forward).

For LegacySerializationPolicy, it is unused when the policy isn't found, e.g. this is an old client with a new server, in which case the old-client-compile-flag is already potentially stale vs. the new-server-config/classes, so I don't think the annotation on the service guarantees the two match in that case either.

It will be interesting to see what your test run against internal Google apps show.

Patch-set: 4