Update patch set 4
Patch Set 4:

re: "The current patch handles that, for StandardSerializationPolicy, by writing the compile-flag into the server-config, which I believe is fine because that's the point of the policy file anyway (recording what the compiler knows for the server to use later)."

Well yes, sort of. The point of the serialization policy file is to say what *not* to send, which is why serialization still works if you swap in a policy that permits anything. Also, most of the data in the serialization policy is a summary of data originally coming from Java classes.  Some is unreadable using regular Java reflection (e.g. super sourcing or GWT module files) so we have to make the GWT compiler do it.

Unfortunately it's a rather fragile data channel; errors due to being unable to find a serialization policy file happen all the time. It's not entirely under our control.

But if we can read a Java annotation directly, that's a more reliable data channel than sending similar data through the GWT compiler and then through the serialization file. Of course this can break too, though things like swapping out jar files, but it seems more reliable in general.

Patch-set: 4