Update patch set 4
Patch Set 4:

The audit versus warn thing seems more like a compiler setting and doesn't belong in an annotation, so let's skip that for now.

Hmm, how about this: we create a new annotation: @GwtRpcSerializable

If a class has this annotation, it doesn't need to extend Serializable or IsSerializable, it doesn't need to be listed in any serialization policy, and it automatically supports serializing final fields.

Then if you're writing a library that contains data transfer objects with GWT in mind, you should be able to write a type with a final field, add the annotation to the class, and have it always be serialized the way you intended, regardless of GWT compiler settings. Also, the class would have a dependency on the annotation, so if you're writing a library that needs to work with an earlier version of GWT, it would be obvious that you can't use this annotation or depend on final field serialization.

So the migration seems pretty simple and painless. How to deal with inheritance less so. Do you need to add this annotation to every subclass you want to transfer? I think I'd say yes, for simplicity, but it's a bit of a pain.

Patch-set: 4