Update patch set 4
Patch Set 4:

Sheesh. Good thing I asked, because at the meetup I thought you were talking about an annotation on the service interface, and not annotations on each of the objects.

I really don't want to add "yet another way to mark your classes as serializable", at least for this meager "all I want is final fields which, wtf, should just work in the first place anyway" use case.

To me, even the fact that this flag exists is a compromise, because it's basically to turn on "bug compatibility" mode for some imagined security non-scenario.

So, no, I'm not going to invest effort in an annotation-based solution.

My snide comment of "well, it's still not going to be landed in 6 months" is starting to ring true.

So, I think you have two options:

1) Let me do some minor clean up (change to a config prop was a good point) and +1 the change as "good enough".

2) -1 the flag-based change, and do the work yourself.

Which do you prefer?

Patch-set: 4