Update patch set 1
Patch Set 1:

> Awesome!
 > 
 > (re. the build failure, it looks like the compiler doesn't find the
 > super sources; I can't understand why just looking at the changes
 > here)

I was aware of the breakage before pushing. I played with it but I think somehow gwtar isn't respecting the incl. and excl. patterns.  It is probably a bug or at most we will need to update the build.xml. On the other hand, incremental and monolithic compile both works ok with the change.

Patch-set: 1