Update patch set 4
Patch Set 4:

Sorry to drag this out. To clarify, I'm not asking you to implement the annotation yet; I think we're still in brainstorming mode. I did talk about putting the annotation on the service interface at the meetup. We can talk, I hope.

I do think it's important to think through how we're going to do this migration and it's worth discussing possible problems and how to solve them. It's unfortunate that GWT-RPC didn't support this from the beginning, but the fact is, it didn't. I've found that wishing away migration problems just results in breakage that we have to fix anyway.

Of course if the problem doesn't actually exist (even though logically it could) then we don't have to solve it. So I think the next step is for me to patch this in and see what the damage is. (It might take a few days since our global presubmit queue is having issues.)

Patch-set: 4