Update patch set 1
Patch Set 1: Code-Review-2

This is not a bug, but needs to be declared this way or GWT will not work.

This is a trick to shorten the emitted code, by using the underscore for the prototype of the current object.

Patch-set: 1
Reviewer: Gerrit User 5385 <5385@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=-2