Update patch set 1
Patch Set 1:

I should have investigate that deeper.

Anyway, as for now it prohibits the usage (e.g.) of undescore library in user js scripts if one uses DirectInstallLinker (I address this problem in my patch to the linker). It's the same problem as with jQuery usage in gwt.

@Daniel, should '_' be added to JsRootScope.COMMON_BUILTINS then?

Patch-set: 1