Update patch set 1
Patch Set 1:

Just as I was happy/relieved that we could let the embedded Jetty and its classloader issues die a good death…

IMO, we should talk about that class loading behavior before or real-soon after this change gets merged. I'd like classloading to be much simpler for SuperDevMode, even if that means not reflecting a production deployment. We could either add a switch to the DevMode to turn that new behavior on/off, or we could change it altogether.

Patch-set: 1
Label: Verified=0