Update patch set 2
Patch Set 2:

> Just as I was happy/relieved that we could let the embedded Jetty
 > and its classloader issues die a good death…

In the GWT meetup  we talked jus to do a small change to what we have right now in order to give the user the same user experience with superdev than devmode.

But I agree, in the breaking release (3.0.0 ?) it's nice to get rid of embedded jetty in favor of any server configured in the IDE, and launch dev or superdev codeservers in someway.

 > 
 > IMO, we should talk about that class loading behavior before or
 > real-soon after this change gets merged. I'd like classloading to
 > be much simpler for SuperDevMode, even if that means not reflecting
 > a production deployment.

I would not to do that for the 2.7.0 but 3.0


 >  We could either add a switch to the
 > DevMode to turn that new behavior on/off, or we could change it
 > altogether.


At the beginning I were playing modifying DevMode. The problem is that we have to make dev.jar depend on codeserver.jar so finally I went to a new class in the codeserver package. But if there is an easy way I would go with it instead of having two classes.

Patch-set: 2