Update patch set 2
Patch Set 2:

> > Just as I was happy/relieved that we could let the embedded Jetty
 > > and its classloader issues die a good death…
 > 
 > In the GWT meetup  we talked jus to do a small change to what we
 > have right now in order to give the user the same user experience
 > with superdev than devmode.
 > 
 > But I agree, in the breaking release (3.0.0 ?) it's nice to get rid
 > of embedded jetty in favor of any server configured in the IDE, and
 > launch dev or superdev codeservers in someway.
 > 
 > >
 > > IMO, we should talk about that class loading behavior before or
 > > real-soon after this change gets merged. I'd like classloading to
 > > be much simpler for SuperDevMode, even if that means not
 > reflecting
 > > a production deployment.
 > 
 > I would not to do that for the 2.7.0 but 3.0

Actually, I would just use the standard classloading strategy for webapps: if your deps are in WEB-INF/lib they'll be used, otherwise the classpath will be used (as a delegation, rather than a "copy to the webapp classloader" as it's done today). There's a risk of regressions for libs that walk the classpath (or use ServiceLoader or otherwise use ClassLoader#getResources()) but they're generally easy to solve (remove the lib from your WEB-INF/lib, or from the classpath; and if that doesn't work, then use an external server).

We should discuss this in the gwt-contrib group though, not here.

 > >  We could either add a switch to the
 > > DevMode to turn that new behavior on/off, or we could change it
 > > altogether.
 > 
 > 
 > At the beginning I were playing modifying DevMode. The problem is
 > that we have to make dev.jar depend on codeserver.jar so finally I
 > went to a new class in the codeserver package. But if there is an
 > easy way I would go with it instead of having two classes.

I was talking about the classloading behavior for server-side classes, not the DevMode vs. SuperDevMode one ;-)

Patch-set: 2
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0