Update patch set 1
Patch Set 1:

Nice job. I like this method, which would create a nice transition from devmode arg pattern with server initialization. Although in the future, I think a transition to using facets in the IDE would be more favorable for servlet container/web app server, that is make the server responsibility move to another plugin/facet. I think once we add in this, we can then move to adding a GWT facet for GPE as well, which would load the GWT sdk jars into the runtime classpath.

Patch-set: 1
Label: Verified=0