Update patch set 3
Patch Set 3:

I think this discussion is getting a bit messy because we are mixing here different issues out of the objetive of this patch. 

Let me summarise why this patch is here: It is the consequence of  the discussion in the GWT-meetup to encourage users to utilize the superdevmode, not widely used because most users still prefer devmode mainly because it is easier to run.  In order to give them a better user experience with superdevmode we thought to make a smooth transition, allowing them to run superdevmode+jetty in one click like they do with devmode+jetty.

So the easier way we saw was to modify a little bit the gwt host launcher (DevMode.class), in order to  provide support for superdevmode in any IDE (not only eclipse), as well as make trivial modify maven plugin, ant tasks, user launchers, etc.

The first idea was just to modify the DevMode.class adding a flag to switch dev/superdev, but the problem is that it will create circular dependencies between dev.jar and codeserver.jar. So instead of using a flag to switch I opted for using a different classname but with the same arguments. Also we needed a special nocache.js file for loading the app.

This patch has these goals:
* The only thing needed to switch to superdevmode is just to change the classname in the launcher 
* It fully prepares the environment providing a stub.nocache.js which loads the codeserver.nocache.js from the correct host:port  without needing to install bookmarkers.
* It's NOT needed big changes in IDE's but changing the classname.
* Everything works as before: integration with IDE ui, java UI, etc.
* If the user wants to use its own server, both superdev and dev modes work out-of-the-box setting the -noserver flag because of the stub.nocache.js 

I think all discussions related with removing JettyLauncher,  improving it,  or providing servlets to run dev or superdevmode etc, could be done in the mailing list or in a dev meeting.

Just let me know if makes sense to continue with this patch (it facilitates the work we are doing in GPE), or to work in other way.

Patch-set: 3