Update patch set 5
Patch Set 5:

> It seems like you already got pretty far at adding a new command.
 > I've been meaning to patch it in and see what the flags look like,
 > but I think it's a good idea to continue with this path.
 > 
 > Has anyone else tried it out yet? What does the help look like?

It has exactly the same parameters than DevMode. It reuses bindAddress logLevel sourceLevel workDir to pass them to CodeServer and JettyLauncher, and uses codeServerPoet for setting the codeserver port.

I shared some screenshots including the help of the command and how it works in this document

https://docs.google.com/presentation/d/1T4RLzdtG84KkatyO7heTYLWlf1QqQwNZxrdQBuTrCuI/edit#slide=id.p

Patch-set: 5
Label: Verified=0