Update patch set 5
Patch Set 5:

It might mean that some tools work with almost no changes, so long as they have a way to pass in a new flag to turn Super Dev Mode on. On the other hand, changing the main class might not be any harder than adding a flag?

It might help to know which tools we are talking about, so we can test this.

Patch-set: 5