Update patch set 5
Patch Set 5:

Making JUnitShell run Super Dev Mode would be an interesting idea if it works. Could running tests with Super Dev Mode be that easy? Would it be better than using -web?

I'm a bit skeptical. It seems like we need to find all the test modules and tell the code server about them, and it would need to understand how test modules are "special". Also, we need to change JUnitShell shell anyway to add the new flag.

Patch-set: 5