Update patch set 5
Patch Set 5:

> Making JUnitShell run Super Dev Mode would be an interesting idea
 > if it works. Could running tests with Super Dev Mode be that easy?

If the new launcher and existing launcher are similar enough, why not?

 > Would it be better than using -web?

This will very least ensures that everything (including our own libs) plays well with superdevmode.
It will also get all the benefits that we are planning for the superdev incremental compile for free; especially if provide an interactive runstyle similar to testgal, it will be a huge performance boost.

 > 
 > I'm a bit skeptical. It seems like we need to find all the test
 > modules and tell the code server about them, and it would need to
 > understand how test modules are "special". 

I'm not sure if I'm following. How it is different than running regular devmode under the hood?

Anyway regardless, we will not lose anything if it doesn't work or ends up requiring some very complicated trickery to make it work.

 > Also, we need to change
 > JUnitShell shell anyway to add the new flag.

I don't know if this really counts as a work but yes you probably need add one line of code :)

Patch-set: 5